requirements.txt
clickhouse-connect>=0.6.1
lamini>=0.10.0
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.3.5